fix(acsl): make specifications parse with vanilla Frama-C
Frama-C 33 rejected eight files. All fixes are syntactic; no contract
changes semantics.

- drop 'requires \typeof(x) <: \type(T)' (13 clauses): WP has no
  \typeof/<:, and the neighbouring \valid_read cast already says it
- cast the recursive logic bodies of match_string, real_len, strncmp and
  sysfs_strlen to their declared return type, not integer
- ghost code is C and cannot call a logic function: replace the four
  'ghost size_t n = strlen(...)' bindings with strlen{Pre}(...)
- strlcat: 'ghost Mid:' is a label without a statement
- expand __visible to nothing under Frama-C, which does not know
  __externally_visible__

All 44 src/*.c now typecheck with no warnings.

+++ src/match_string.h
@@ -24,13 +24,13 @@
 
     logic size_t match_string(char **a, size_t n, char *s) =
        n == 0 ?
-          0
+          (size_t)0
        :
-          (strcmp(a[0], s) == 0 ? 0 : match_string(a + 1, (size_t)(n - 1), s));
+          (strcmp(a[0], s) == 0 ? (size_t)0 : match_string(a + 1, (size_t)(n - 1), s));
 
     logic size_t real_len(char **a, size_t n) =
        ((a[0] == NULL) || (n == 0)) ?
-          0
+          (size_t)0
        :
           (size_t)(1 + real_len(a + 1, (size_t)(n - 1)));
